requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
ffmpeg==1.4
google-api-core==2.5.0
google-api-python-client==2.38.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
jedi==0.18.1
moviepy==1.0.3
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
protobuf==3.19.4
pudb==2022.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.7
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.8
urwid==2.1.2
urwid-readline==0.13
